+++ src/muya/lib/contentState/inputCtrl.js
@@ -43,6 +43,9 @@ const inputCtrl = ContentState => {
 
   ContentState.prototype.inputHandler = function (event) {
     const { start, end } = selection.getCursorRange()
+    if (!start || !end) {
+      return
+    }
     const { start: oldStart, end: oldEnd } = this.cursor
     const key = start.key
     const block = this.getBlock(key)