+++ src/muya/lib/selection/index.js
@@ -438,6 +438,15 @@ class Selection {
 
   getCursorRange () {
     let { anchorNode, anchorOffset, focusNode, focusOffset } = this.doc.getSelection()
+    let startParagraph = findNearestParagraph(anchorNode)
+    let endParagraph = findNearestParagraph(focusNode)
+    if (!startParagraph || !endParagraph) {
+      return {
+        start: null,
+        end: null
+      }
+    }
+
     // when the first paragraph is task list, press ctrl + a, then press backspace will cause bug
     // use code bellow to fix the bug
     const findFirstTextNode = anchor => {
@@ -452,9 +461,6 @@ class Selection {
     if (anchorNode.nodeName === 'LI') {
       anchorNode = findFirstTextNode(anchorNode)
     }
-
-    let startParagraph = findNearestParagraph(anchorNode)
-    let endParagraph = findNearestParagraph(focusNode)
 
     const getOffsetOfParagraph = (node, paragraph) => {
       let offset = 0