+++ .github/CHANGELOG.md
@@ -86,7 +86,8 @@ This update **fixes a XSS security vulnerability** when exporting a document.
 - Fixed an issue that relative images are not loaded after closing a tab
 - Add symbolic link support
 - Fixed bug when combine pre list and next list into one when inline update #707
-- Fixed list parse error https://github.com/marktext/marktext/issues/831#issuecomment-477719256
+- Fix renderer error when selection in sidebar (#625)
+- Fixed list parse error [more info](https://github.com/marktext/marktext/issues/831#issuecomment-477719256)
 
 ### 0.13.65